requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.2
azure-identity==1.17.1
azure-storage-blob==12.20.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.140
botocore==1.34.140
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.1
fastapi==0.109.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
genson==1.3.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.26.0
huggingface-hub==0.23.4
hyperframe==6.0.1
idna==3.7
indexify==0.0.29
indexify-extractor-sdk==0.0.84
isodate==0.6.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-text-splitters==0.2.2
langsmith==0.1.83
llama-cloud==0.0.6
llama-index==0.10.52
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.52.post1
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.2.3
llama-index-legacy==0.9.48
llama-index-llms-mistralai==0.1.17
llama-index-llms-openai==0.1.25
llama-index-multi-modal-llms-openai==0.1.7
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.27
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mistralai==0.4.2
mpmath==1.3.0
msal==1.29.0
msal-extensions==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
nanoid==2.0.0
nest-asyncio==1.6.0
netifaces2==0.0.21
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
openai==1.35.10
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pillow==10.4.0
portalocker==2.10.0
proto-plus==1.24.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3fs==0.4.2
s3transfer==0.10.2
safetensors==0.4.3
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.36.3
streamlit==1.36.0
striprtf==0.0.26
sympy==1.12.1
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.42.3
triton==2.3.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.27.1
watchdog==4.0.1
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
